+++ src/components/about/About.css
@@ -1,13 +1,7 @@
 .about-container {
   display: flex;
-
-  background: rgb(41, 41, 41);
-  background: linear-gradient(
-    90deg,
-    rgb(124, 19, 19) 0%,
-    rgb(4, 173, 142) 100%
-  );
-  color: white;
+  background: linear-gradient(100deg, rgb(19, 62, 173) 0%, rgb(0, 0, 0) 60%);
+  color: rgb(251, 244, 244);
   height: 70vh;
   padding: 100px, 0, 100px, 0;
 }
@@ -16,15 +10,15 @@
   display: flex;
   flex-direction: column;
   justify-content: center;
-  width: 55%;
-  padding: 0 100px 0 100px;
+  width: 60%;
+  padding: 10px 50px 10px 100px;
   font-family: "Nanum Gothic", sans-serif;
 }
 
 .about-description > h3 {
   margin-bottom: 30px;
   font-size: 28px;
-  color: white;
+  color: rgb(255, 187, 187);
   text-shadow: 2px 2px 4px #000000;
   font-family: "Nanum Gothic", sans-serif;
 }
@@ -36,15 +30,17 @@
 }
 
 .about-img {
-  background-image: conic-gradient(red 45deg, yellow 90deg, green 210deg);
-  height: 80%;
-  width: 28%;
+  height: 70%;
+  width: 19%;
   display: flex;
   align-self: center;
   overflow: hidden;
   border-radius: 50%;
-  border-style: ridge;
-  box-shadow: 0 0 20px rgb(0, 0, 0);
+  box-shadow: 0 0 50px rgb(0, 0, 135);
+}
+.about-img > img {
+  width: 100%;
+  height: 100%;
 }
 
 @media only screen and (max-width: 960px) {
@@ -55,7 +51,7 @@
   }
 
   .about-desc {
-    width: 90%;
+    width: 20%;
     margin: 0 auto;
     padding: 40px 20px 50px 20px;
   }
@@ -66,7 +62,7 @@
   }
 
   .about-img {
-    width: 100%;
+    width: 50%;
     margin-bottom: 50px;
   }
