+++ src/components/cover/Cover.css
@@ -21,16 +21,16 @@
   font-size: 150px;
   margin: 0 auto;
   text-align: center;
-  text-shadow: -3px 3px 50px white;
+  text-shadow: 2px 2px 5px #03f82c;
   font-family: "Inspiration", cursive;
 }
 
 .cover-container > p {
-  color: white;
+  color: rgb(255, 217, 27);
   padding-left: 25%;
   margin-top: 10px;
   font-size: 25px;
-  text-shadow: -3px 4px 4px #000000;
+  text-shadow: 2px 2px 5px #000000;
   font-family: "Rubik Iso", cursive;
 }
