+++ src/components/footer/Footer.css
@@ -1,8 +1,8 @@
 .footer {
   width: 100%;
   height: 100vh;
-  background-color: #292929;
-  color: white;
+  background: linear-gradient(100deg, rgb(19, 62, 173) 0%, rgb(0, 0, 0) 60%);
+  color: rgb(176, 173, 173);
   padding-top: 180px;
   position: fixed;
   z-index: -2;
@@ -11,12 +11,12 @@
 }
 
 .footer-info {
-  font-size: 35px;
+  font-size: 50px;
   padding: 30px 0 30px 50px;
   width: 70%;
   margin: 0 auto;
   text-align: center;
-  text-shadow: 2px 2px 4px #ffffff;
+  text-shadow: 2px 2px 4px #090808;
   font-family: "Nanum Gothic", sans-serif;
 }
 
@@ -24,24 +24,34 @@
   padding-bottom: 15px;
   text-align: center;
   text-shadow: 1px 2px;
-  text-shadow: 4px 5px 4px #8a8a15;
+  text-shadow: 4px 2px 5px #d3f300;
 }
 
 .footer-contact {
   font-size: 22px;
-  padding: 200px 0 30px 50px;
-  width: 70%;
+  padding: 115px 10px 50px;
+  width: 50%;
   margin: 0 auto;
   text-shadow: 1px 2px 5px #8a8a15;
   font-family: "Nanum Gothic", sans-serif;
 }
 
+.footer-contact > p {
+  color: #000000;
+  text-shadow: 2px 2px 4px #ffffff;
+}
+.footer-contact > h3 {
+  height: 20px;
+  color: #000000;
+  text-shadow: 2px 2px 4px #ffffff;
+}
+
 .footer-sns {
-  width: 70%;
+  width: 80%;
   line-height: 1.5;
-  text-shadow: 2px 2px 5px #37e657;
+  text-shadow: 2px 2px 5px #849bda;
   position: absolute;
-  bottom: 40px;
+  bottom: 90px;
   left: 0;
   right: 0;
   margin: 0 auto;
@@ -52,13 +62,13 @@
 
 .sns-link {
   display: grid;
-  width: 100%;
+  width: 80%;
   grid-template-columns: 1fr 1fr 1fr;
 }
 
 .linkedin,
 .github {
-  font-size: 35px;
+  font-size: 40px;
   margin: 0 auto;
   padding-bottom: 10px;
   border-bottom: 4px solid transparent;
@@ -66,7 +76,7 @@
 
 .linkedin:hover,
 .github:hover {
-  border-color: #dae637;
+  border-color: #04a0e8;
   transition: all 0.3s ease-in-out;
 }
