+++ src/components/info/Info.css
@@ -1,9 +1,14 @@
 .info-container {
   width: 100%;
   height: 300px;
-  background: white;
+  background: linear-gradient(
+    100deg,
+    rgb(29, 29, 29) 34%,
+    rgb(223, 176, 4) 80%
+  );
   color: black;
   margin-bottom: 100vh;
+  font-size: 25px;
 }
 
 .info {
@@ -11,7 +16,7 @@
   justify-content: center;
   align-items: center;
   height: 100%;
-  text-shadow: 0 0 3px #ff0000;
+  text-shadow: 0 0 3px #fcf9f9;
   font-family: "Nanum Gothic", sans-serif;
 }
